Move test to more appropriate test file; include backspace counterpart
