+++ richtextfx/src/integrationTest/java/org/fxmisc/richtext/keyboard/DeletionTests.java
@@ -3,6 +3,7 @@ package org.fxmisc.richtext.keyboard;
 import com.nitorcreations.junit.runners.NestedRunner;
 import javafx.stage.Stage;
 import org.fxmisc.richtext.InlineCssTextAreaAppTest;
+import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 
@@ -94,5 +95,36 @@ public class DeletionTests {
             assertEquals(pos - 1, area.getCaretPosition());
         }
 
+    }
+
+    // miscellaneous cases
+
+    public class When_Area_Ends_With_Empty_Line extends InlineCssTextAreaAppTest {
+
+        @Override
+        public void start(Stage stage) throws Exception {
+            super.start(stage);
+            area.replaceText(0, 0, "abc\n");
+        }
+
+        public class And_All_Text_Is_Selected {
+
+            @Before
+            public void selectAllText() {
+                interact(() -> area.selectAll());
+            }
+
+
+            @Test
+            public void pressing_delete_should_not_throw_exception() {
+                push(DELETE);
+            }
+
+            @Test
+            public void pressing_backspace_should_not_throw_exceptions() {
+                push(BACK_SPACE);
+            }
+
+        }
     }
 }
